+++ minimal_uhdless_demo/src/time_spec.cpp
@@ -42,17 +42,17 @@ time_spec_t::time_spec_t(int64_t full_secs, double frac_secs)
 
 time_spec_t::time_spec_t(int64_t full_secs, long tick_count, double tick_rate)
 {
-    const double frac_secs = tick_count / tick_rate;
+    const double frac_secs = static_cast<double>(tick_count) / tick_rate;
     time_spec_init(full_secs, frac_secs);
 }
 
 time_spec_t time_spec_t::from_ticks(long long ticks, double tick_rate)
 {
     const long long rate_i      = (long long)(tick_rate);
-    const double rate_f         = tick_rate - rate_i;
+    const double rate_f         = tick_rate - static_cast<double>(rate_i);
     const int64_t secs_full     = int64_t(ticks / rate_i);
     const long long ticks_error = ticks - (secs_full * rate_i);
-    const double ticks_frac     = ticks_error - secs_full * rate_f;
+    const double ticks_frac     = static_cast<double>(ticks_error) - static_cast<double>(secs_full) * rate_f;
     return time_spec_t(secs_full, ticks_frac / tick_rate);
 }
 
@@ -67,16 +67,16 @@ long time_spec_t::get_tick_count(double tick_rate) const
 long long time_spec_t::to_ticks(double tick_rate) const
 {
     const long long rate_i     = (long long)(tick_rate);
-    const double rate_f        = tick_rate - rate_i;
+    const double rate_f        = tick_rate - static_cast<double>(rate_i);
     const long long ticks_full = this->get_full_secs() * rate_i;
-    const double ticks_error   = this->get_full_secs() * rate_f;
+    const double ticks_error   = static_cast<double>(this->get_full_secs()) * rate_f;
     const double ticks_frac    = this->get_frac_secs() * tick_rate;
     return ticks_full + fast_llround(ticks_error + ticks_frac);
 }
 
 double time_spec_t::get_real_secs(void) const
 {
-    return this->get_full_secs() + this->get_frac_secs();
+    return static_cast<double>(this->get_full_secs()) + this->get_frac_secs();
 }
 
 void time_spec_t::sleep_for(void) const {
@@ -100,7 +100,7 @@ time_spec_t& time_spec_t::operator+=(double& rhs)
 {
     double full_secs = std::trunc(rhs);
     time_spec_init(
-        this->get_full_secs() + full_secs, this->get_frac_secs() + rhs - full_secs);
+        static_cast<double>(this->get_full_secs()) + full_secs, this->get_frac_secs() + rhs - full_secs);
     return *this;
 }
 
@@ -115,7 +115,7 @@ time_spec_t& time_spec_t::operator-=(double& rhs)
 {
     double full_secs = std::trunc(rhs);
     time_spec_init(
-        this->get_full_secs() - full_secs, this->get_frac_secs() - (rhs - full_secs));
+        static_cast<double>(this->get_full_secs()) - full_secs, this->get_frac_secs() - (rhs - full_secs));
     return *this;
 }
